requirements.txt
streamlit
streamlit-chat
openai
langchain-groq
langchain
langchain-openai
langchain-community
langchain-chroma
langchainhub
beautifulsoup4
requests
python-dotenv